Safari-611.1.10/LayoutTests/accessibility/mac/aria-radiogroup-expected.txt

radio 1
radio 2
This tests that the aria radiogroup role is correctly mapped to a WebCore accessibility role.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
